psr-15 wrappers code improvement
