--- src/InstantiatingRouterInterface.php
@@ -1,41 +1,0 @@
-<?php
-//
-// +---------------------------------------------------------------------+
-// | CODE INC. SOURCE CODE                                               |
-// +---------------------------------------------------------------------+
-// | Copyright (c) 2018 - Code Inc. SAS - All Rights Reserved.           |
-// | Visit https://www.codeinc.fr for more information about licensing.  |
-// +---------------------------------------------------------------------+
-// | NOTICE:  All information contained herein is, and remains the       |
-// | property of Code Inc. SAS. The intellectual and technical concepts  |
-// | contained herein are proprietary to Code Inc. SAS are protected by  |
-// | trade secret or copyright law. Dissemination of this information or |
-// | reproduction of this material is strictly forbidden unless prior    |
-// | written permission is obtained from Code Inc. SAS.                  |
-// +---------------------------------------------------------------------+
-//
-// Author:   Joan Fabrégat <joan@codeinc.fr>
-// Date:     27/09/2018
-// Project:  Router
-//
-declare(strict_types=1);
-namespace CodeInc\Router;
-use Psr\Http\Message\ServerRequestInterface;
-
-
-/**
- * Interface InstantiatingRouterInterface
- *
- * @package CodeInc\Router
- * @author Joan Fabrégat <joan@codeinc.fr>
- */
-interface InstantiatingRouterInterface extends RouterInterface
-{
-    /**
-     * Returns an instance of the controller to handle the given HTTP request or NULL if none is available.
-     *
-     * @param ServerRequestInterface $request
-     * @return ControllerInterface|null
-     */
-    public function getController(ServerRequestInterface $request):?ControllerInterface;
-}
